e2e: log number of sent txs (success and failed)
+

+++ test/e2e/runner/load.go
@@ -70,7 +70,7 @@ func Load(ctx context.Context, testnet *e2e.Testnet) error {
 		// Log every ~1 second the number of sent transactions.
 		total := success + failed
 		if total%testnet.LoadTxBatchSize == 0 {
-			logger.Debug("load", "success", success, "failed", failed, "success/total", log.NewLazySprintf("%.1f", success/total), "tx/s", rate)
+			logger.Debug("load", "success", success, "failed", failed, "success/total", log.NewLazySprintf("%.1f", float64(success)/float64(total)), "tx/s", rate)
 		}
 
 		// Check if reached max number of allowed transactions to send.
